File structure refactored to fractals
