+++ src/components/loginRegisterModal/RegisterForm.tsx
@@ -1,8 +1,8 @@
 import { Field, FormikHelpers, Formik } from 'formik';
 import React from 'react';
 import * as yup from 'yup';
-import { RegisterFormValues } from '../types';
-import FormField from './FormField';
+import { RegisterFormValues } from '../../types';
+import FormField from '../FormField';
 
 type Props = {
     switchToLogin: (values: RegisterFormValues) => void;
