+++ src/components/navigationBar/MobileMenuButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { ReactComponent as MenuIcon } from '../images/icons/menu.svg';
-import { ReactComponent as CloseIcon } from '../images/icons/close.svg';
+import { ReactComponent as MenuIcon } from '../../images/icons/menu.svg';
+import { ReactComponent as CloseIcon } from '../../images/icons/close.svg';
 
 type Props = {
     onClick: () => void;
