+++ src/pages/Home.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
+import LoginRegisterModal from '../components/LoginRegisterModal';
 
 function Home() {
     return (
         <div>
-            <p>Under Construction</p>
+            <LoginRegisterModal />
         </div>
     );
 }
